refactor: 이름 변경에 따른 변경
