--- src/main/java/com/d108/project/domain/common/enums/AgeGroup.java
@@ -1,5 +1,0 @@
-package com.d108.project.domain.common.enums;
-
-public enum AgeGroup {
-
-}
