+++ src/main/java/com/d108/project/domain/forum/post/service/PostService.java
@@ -1,0 +1,20 @@
+package com.d108.project.domain.forum.post.service;
+
+import com.d108.project.domain.forum.post.dto.PostUpdateDto;
+import com.d108.project.domain.forum.post.dto.PostCreateDto;
+import com.d108.project.domain.forum.post.dto.PostResponseDto;
+
+import java.util.List;
+
+public interface PostService {
+
+    void createPost(PostCreateDto postCreateDto);
+
+    List<PostResponseDto> getAllPosts();
+
+    PostResponseDto getPostById(Integer id);
+
+    void updatePostById(Integer id, PostUpdateDto postUpdateDto);
+
+    void deletePostById(Integer id);
+}
