+++ src/main/java/com/d108/project/domain/forum/reply/entity/Reply.java
@@ -1,0 +1,26 @@
+package com.d108.project.domain.forum.reply.entity;
+
+import com.d108.project.domain.global.BaseTimeEntity;
+import com.d108.project.domain.forum.post.entity.Post;
+import com.d108.project.domain.member.entity.Member;
+import jakarta.persistence.*;
+import lombok.Getter;
+
+@Entity
+@Getter
+@Table(name = "replies")
+public class Reply extends BaseTimeEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
+
+    private String content;
+}
