+++ src/main/java/com/d108/project/domain/forum/saleStorePost/repository/SaleStorePostRepository.java
@@ -1,0 +1,7 @@
+package com.d108.project.domain.forum.saleStorePost.repository;
+
+import com.d108.project.domain.forum.saleStorePost.entity.SaleStorePost;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SaleStorePostRepository extends JpaRepository<SaleStorePost, Integer> {
+}
