--- src/main/java/com/d108/project/domain/forum/salestorepost/repository/SaleStorePostRepository.java
@@ -1,7 +1,0 @@
-package com.d108.project.domain.forum.salestorepost.repository;
-
-import com.d108.project.domain.forum.salestorepost.domain.SaleStorePost;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SaleStorePostRepository extends JpaRepository<SaleStorePost, Integer> {
-}
