+++ src/main/java/com/d108/project/domain/global/BaseTimeEntity.java
@@ -1,0 +1,27 @@
+package com.d108.project.domain.global;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.AccessLevel;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@Getter(AccessLevel.PUBLIC)
+@MappedSuperclass // BaseEntity 를 상속한 엔티티들은 아래 필드들을 컬럼으로 인식하게 된다.
+@EntityListeners(AuditingEntityListener.class)
+// Spring Data JPA가 Auditing(자동으로 값 매핑) 기능 추가
+// 도메인을 영속성 컨텍스트에 저장하거나 조회를 수행한 후에 update를 하는 경우,
+public class BaseTimeEntity {
+
+    @CreatedDate
+    @Column(updatable = false)
+    private LocalDateTime createdAt;
+
+    @LastModifiedDate
+    private LocalDateTime updatedAt;
+}
