+++ src/main/java/com/d108/project/domain/global/enums/AgeGroup.java
@@ -1,0 +1,5 @@
+package com.d108.project.domain.global.enums;
+
+public enum AgeGroup {
+
+}
