+++ src/main/java/com/d108/project/domain/logincredential/repository/LoginCredentialRepository.java
@@ -1,6 +1,6 @@
 package com.d108.project.domain.logincredential.repository;
 
-import com.d108.project.domain.logincredential.domain.LoginCredential;
+import com.d108.project.domain.logincredential.entity.LoginCredential;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
